+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Unreleased
+## [Unreleased]
 
 ### Added
 
@@ -42,4 +42,5 @@
 
   * Initial release
 
+[Unreleased]: https://github.com/stjude/fqlib/compare/v0.2.0...HEAD
 [0.2.0]: https://github.com/stjude/fqlib/compare/v0.1.0...v0.2.0