+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+## [0.3.0] - 2019-08-09
+
 ### Added
 
   * New `filter` command. This accepts a whitelist of record read names to keep
@@ -46,5 +48,6 @@
 
   * Initial release
 
-[Unreleased]: https://github.com/stjude/fqlib/compare/v0.2.0...HEAD
+[Unreleased]: https://github.com/stjude/fqlib/compare/v0.3.0...HEAD
+[0.3.0]: https://github.com/stjude/fqlib/compare/v0.2.0...v0.3.0
 [0.2.0]: https://github.com/stjude/fqlib/compare/v0.1.0...v0.2.0